Separate API routes into separate folders.
