+++ routes/common.js
@@ -1,0 +1,3 @@
+export function removeHtmlEntities(str) {
+    return str.replace(/&ldquo;|&rdquo;/g, '');
+}
